backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-environ==0.11.2
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.8.2
requests==2.32.3
